Update server.py
Fixes a problem where if the chat has been going on for long there will be a error when using pickle.loads because there is a EOF because the messages are larger than the sockets buffsize. By checking the different between the last sent messages and the messages on the server. This means only the latest message is sent to the clients